fix(site-page): long view page to last page
When the next chapter does not have as many pages as
the previous chapter, the order of the rendering
makes it so that it does not find the correct
element to scroll to for the last page.

It seems like this fix works fairly consistently,
probably because rendering and the message sending
is likely synchronous all on the main UI thread,
so it finishes the render for the current message
and then starts the next message, which is the
message send.