fix(lib): handle Routing requests
Further context for previous update,
Azure has a setting for rerouting
certain requests when there is a more optimal routing
path, such as when the connection comes
from inside the Azure network.

Tiberius returns this as a Routing Error
with the host and port requested,
so we handle that explicitly and gracefully without user error.

Since it's not connection pooling yet,
this rerouting would happen every time when
entering a non-optimal route, so we leave a warning despite being able to handle it.